check

+++ src/modules/comments/application/usecases/delete-comments.usecase.ts
@@ -2,9 +2,13 @@ import { Types } from 'mongoose';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { Result } from '../../../../core/types/result';
 import { CommentsRepository } from '../../../comments/infrastructure/comments.repository';
+import { ResultStatus } from '../../../../core/types/result-code';
 
 export class DeleteCommentsCommand {
-  constructor(public id: Types.ObjectId) {}
+  constructor(
+    public id: Types.ObjectId,
+    public userId: string,
+  ) {}
 }
 
 @CommandHandler(DeleteCommentsCommand)
@@ -14,7 +18,22 @@ export class DeleteCommentsUseCase implements ICommandHandler<
 > {
   constructor(private commentsRepository: CommentsRepository) {}
 
-  async execute({ id }: DeleteCommentsCommand): Promise<Result<null>> {
+  async execute({ id, userId }: DeleteCommentsCommand): Promise<Result<null>> {
+    const comment = await this.commentsRepository.findCommentById(id);
+
+    if (comment.status !== ResultStatus.Success) {
+      return comment;
+    }
+
+    if (comment.data.commentatorInfo.userId.toString() !== userId) {
+      return {
+        data: null,
+        status: ResultStatus.Forbidden,
+        errorMessage: '',
+        extensions: [],
+      };
+    }
+
     return this.commentsRepository.deleteCommentById(id);
   }
 }
